Adding new MqttHub library which is a better version of the EFM32COMM library
This library supports multiple HUB's which can each have multiple ports
and modules attached. Also supports new MQTT cc3000 library and has
simple configuration of cc3000 and MQTT cc3000. See MQTTTest1 example in
this library to see how to use it.

MqttHub library file missed in the last checkin

Minor changes to the MQTT library to be easier to configure and remove printouts.

Added support for SPI objects other than the default SPI to be used with the CC3000. NOTE: still limited to a single cc3000 chip per app.

Fixed SPI1 to work and reconfigured SPI on WG842 to work on the Garage controller.

Updated MqttHub examples to work with MqttHub instead of EFM32COMM.

Minor Cleanup